+++ tally/spec/settings_spec.lua
@@ -1,6 +1,6 @@
 insulate("Settings parser", function()
 
-    local function mockSettings(fileName)
+    local function mockSettings(fileName, additionalLines)
         -- mocks
         _G.node = {
             chipid = function() return 12345678 end
@@ -10,6 +10,10 @@ insulate("Settings parser", function()
             open = function()
                 local lines = {}
                 for line in io.lines("spec/fixtures/" .. fileName) do
+                    table.insert(lines, line)
+                end
+
+                for _, line in pairs(additionalLines or {}) do
                     table.insert(lines, line)
                 end
 
@@ -83,6 +87,122 @@ insulate("Settings parser", function()
             assert.is_true(MySettings.name():len() <= 32)
             assert.is_true(MySettings.hostName():len() <= 32)
             assert.is_same("Tally-this-is-a-really-messe", MySettings.hostName())
+        end)
+    end)
+
+    insulate("operator.type", function()
+        insulate("should be grb+ by default", function()
+            it(function()
+                mockSettings("settings-all.ini")
+                require "src.my-settings"
+
+                assert.is_same("grb+", MySettings.operatorType())
+            end)
+        end)
+        insulate("should support grb+", function()
+            it(function()
+                mockSettings("settings-all.ini", {
+                    "operator.type=grb+"
+                })
+                require "src.my-settings"
+
+                assert.is_same("grb+", MySettings.operatorType())
+            end)
+        end)
+        insulate("should support grb-", function()
+            it(function()
+                mockSettings("settings-all.ini", {
+                    "operator.type=grb-"
+                })
+                require "src.my-settings"
+
+                assert.is_same("grb-", MySettings.operatorType())
+            end)
+        end)
+        insulate("should ignore case", function()
+            it(function()
+                mockSettings("settings-all.ini", {
+                    "operator.type=GrB-"
+                })
+                require "src.my-settings"
+
+                assert.is_same("grb-", MySettings.operatorType())
+            end)
+        end)
+        insulate("should log a warning on invalid value and use the default", function()
+            it(function()
+                mockSettings("settings-all.ini", {
+                    "operator.type=invalid"
+                })
+
+                local warnings = {}
+                _G.MyLog = {
+                    warning = function(warning) table.insert(warnings, warning) end
+                }
+
+                require "src.my-settings"
+
+                assert.is_same("Invalid operator.type \"invalid\"", warnings[1])
+                assert.is_same("grb+", MySettings.operatorType())
+            end)
+        end)
+    end)
+
+    insulate("stage.type", function()
+        insulate("should be grb+ by default", function()
+            it(function()
+                mockSettings("settings-all.ini")
+                require "src.my-settings"
+
+                assert.is_same("grb+", MySettings.stageType())
+            end)
+        end)
+        insulate("should support grb+", function()
+            it(function()
+                mockSettings("settings-all.ini", {
+                    "stage.type=grb+"
+                })
+                require "src.my-settings"
+
+                assert.is_same("grb+", MySettings.stageType())
+            end)
+        end)
+        insulate("should support grb-", function()
+            it(function()
+                mockSettings("settings-all.ini", {
+                    "stage.type=grb-"
+                })
+                require "src.my-settings"
+
+                assert.is_same("grb-", MySettings.stageType())
+            end)
+        end)
+        insulate("should ignore case", function()
+            it(function()
+                mockSettings("settings-all.ini", {
+                    "stage.type=GrB-"
+                })
+                require "src.my-settings"
+
+                assert.is_same("grb-", MySettings.stageType())
+            end)
+        end)
+        insulate("should log a warning on invalid value and use the default", function()
+            it(function()
+                mockSettings("settings-all.ini", {
+                    "stage.type=invalid"
+                })
+
+                local warnings = {}
+                _G.MyLog = {
+                    warning = function(warning) table.insert(warnings, warning) end
+                }
+
+                require "src.my-settings"
+
+                assert.is_same("Invalid stage.type \"invalid\"", warnings[1])
+                assert.is_same("grb+", MySettings.stageType())
+            end)
         end)
     end)