+++ tally/src/my-settings.lua
@@ -1,14 +1,28 @@
 local fileName = "/FLASH/tally-settings.ini"
 
+_G.LightTypes = {
+    COMMON_ANODE = "grb+",
+    COMMON_CATHODE = "grb-",
+}
+
 -- defaults
 local staSsid = nil
 local staPw = nil
 local hubIp = nil
 local hubPort = 7411
 local name = string.format("%x", node.chipid())
+local operatorType = LightTypes.COMMON_ANODE
+local stageType = LightTypes.COMMON_ANODE
 
 local trim = function(s)
     return s:match("^%s*(.-)%s*$")
+end
+
+local isValueInTable = function(value, table)
+    for _, val in pairs(table) do
+        if value == val then return true end
+    end
+    return false
 end
 
 local makeHostName = function(s)
@@ -33,6 +47,12 @@ _G.MySettings = {
     end,
     hostName = function()
         return makeHostName(name)
+    end,
+    operatorType = function()
+        return operatorType
+    end,
+    stageType = function()
+        return stageType
     end,
     isValid = function()
         return staSsid ~= nil and hubIp ~= nil
@@ -61,6 +81,20 @@ if file.exists(fileName) then
                         hubPort = tonumber(v)
                     elseif k == "tally.name" then
                         name = v:sub(0, 26)
+                    elseif k == "operator.type" then
+                        local value = v:lower()
+                        if isValueInTable(value, LightTypes) then
+                            operatorType = value
+                        else
+                            MyLog.warning("Invalid operator.type \"" .. v .. "\"")
+                        end
+                    elseif k == "stage.type" then
+                        local value = v:lower()
+                        if isValueInTable(value, LightTypes) then
+                            stageType = value
+                        else
+                            MyLog.warning("Invalid stage.type \"" .. v .. "\"")
+                        end
                     else
                         MyLog.warning("Unknown key " .. k .. " in configuration file. Did you mistype it?")
                     end